moved outdated cleaning logic to dedicated hook
