+++ app/src/hooks/useOngoingTransfersTracker.tsx
@@ -7,13 +7,12 @@ import {
   TxTrackingResult,
 } from '@/models/transfer'
 import { resolveDirection } from '@/services/transfer'
-import { getExplorerLink, startedTooLongAgo } from '@/utils/transfer'
+import { getExplorerLink } from '@/utils/transfer'
 import {
   findMatchingTransfer,
   getTransferStatus,
   isCompletedTransfer,
 } from '@/utils/transferTracking'
-import { captureException } from '@sentry/nextjs'
 import { TransferStatus } from '@snowbridge/api/dist/history'
 import { useCallback, useEffect, useRef, useState } from 'react'
 import useCompletedTransfers from './useCompletedTransfers'
@@ -155,37 +154,6 @@ const useOngoingTransfersTracker = () => {
       }
     })
   }, [transfers, addCompletedTransfer, remove, ongoingTransfers, addNotification, updateUniqueId])
-
-  useEffect(() => {
-    ongoingTransfers.forEach(ongoing => {
-      if (startedTooLongAgo(ongoing)) {
-        const explorerLink = getExplorerLink(ongoing)
-        remove(ongoing.id)
-        addCompletedTransfer({
-          id: ongoing.id,
-          result: TxStatus.Undefined,
-          token: ongoing.token,
-          sourceChain: ongoing.sourceChain,
-          destChain: ongoing.destChain,
-          amount: ongoing.amount,
-          tokenUSDValue: ongoing.tokenUSDValue ?? 0,
-          fees: ongoing.fees,
-          sender: ongoing.sender,
-          recipient: ongoing.recipient,
-          date: ongoing.date,
-          ...(explorerLink && { explorerLink }),
-        } satisfies CompletedTransfer)
-
-        addNotification({
-          message: 'Transfer verification failed.',
-          severity: NotificationSeverity.Warning,
-          dismissible: true,
-        })
-        captureException(new Error('Transfer tracking failed'), { extra: { ongoing } })
-      }
-    })
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [ongoingTransfers])
 
   return { transfers, loading, statusMessages, fetchTransfers }
 }
